Enhancement: Add project top feature component Finished dropdown action Re #370
