+++ app/ui/projects/detail/route.js
@@ -1,19 +1,9 @@
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
-import { computed, get } from '@ember/object';
+import { computed } from '@ember/object';
 import { readOnly } from '@ember/object/computed';
 
-function findByComponentNameAndPath(model, name, path) {
-	if (model && model.sections) {
-		return get(model.sections.find(s => s.component === name), path);
-	}
-
-	return null;
-}
-
 export default Route.extend({
-	queryParams: 'topfeatures',
-
 	staticContent: service(),
 	win: service(),
 
@@ -28,25 +18,6 @@ export default Route.extend({
 			return acc;
 		}, []);
 	}),
-
-	topfeatures: computed(function() {
-		return findByComponentNameAndPath(
-			this.get('staticContent').readShoebox('/index'),
-			'project-features',
-			'features.0.id'
-		);
-	}),
-
-	actions: {
-		scrollToSection(event) {
-			event.preventDefault();
-
-			this.win.scrollToAnimated(
-				window.pageYOffset +
-					document.querySelector(event.target.hash).getBoundingClientRect().top
-			);
-		},
-	},
 
 	titleToken(model) {
 		return model.meta.title;
