+++ app/ui/projects/detail/template.hbs
@@ -14,7 +14,7 @@
 	{{#project-detail data=model}}
 		{{#each model.sections as |data|}}
 			{{#if data.features}}
-				{{component data.component data=data selectedId=feature-one}}
+				{{component data.component data=data selected=topfeatures}}
 			{{else}}
 				{{component data.component data=data meta=model.meta}}
 			{{/if}}
